Update Board and Main in Minesweeper
Changed character for mine representation in the Board and added a user input for the number of mines in the Main. This allows more interactivity, letting the user decide how many mines to place on the field when playing Minesweeper. A validation loop is also included to ensure input is a number.